Update Makefile and Cargo.toml and docs to refer to porky only
