+++ porky/BUILDING.md
@@ -15,17 +15,19 @@ cd pigg/porky
 make
 ```
 
-will run build a release DWARF binary for the Raspberry Pi Pico W here:
+will run build a release DWARF binary for the Raspberry Pi Pico, Pico W, Pico 2 and Pico 2 W:
 
-- `pigg/porky/target/thumbv6m-none-eabi/release/porky_w`
+The same binary is overwritten for each, the last build being Pi Pico 2 W, which can be found at:
 
-### How SSID information is used by `porky_w`
+- `pigg/porky/target/thumbv6m-none-eabi/release/porky`
 
-The startup process of `porky_w` is:
+### How SSID information is used by `porky` on Pico W/Pico 2 W
+
+The startup process of `porky` is:
 
 - Try to load SSID information from Flash memory
 - If none is present try and use the default SSID information compiled in as part of the build
-- If no default information was built in, then `porky_w` has no SSID information and will not try to connect to a Wi-Fi
+- If no default information was built in, then `porky` has no SSID information and will not try to connect to a Wi-Fi
   network
 - USB connection is always started and will be able to receive SSID information from `piggui`. This will be stored in
   the Flash memory and used in the above process at next start
@@ -39,7 +41,7 @@ of how to do it using `piggui`
 
 ### Customizing the default SSID information
 
-It is possible to customize the default SSID (Wi-Fi network) information that `porky_w` tries to use to connect to a
+It is possible to customize the default SSID (Wi-Fi network) information that `porky` tries to use to connect to a
 Wi-Fi network on startup, if you are building from source. Then all devices programmed with that build will
 automatically try to connect to that Wi-Fi network at startup, without having to use `piggui` and USB.
 
@@ -56,24 +58,30 @@ security = "wpa2"
 
 Valid values for the `security` field are: `open`, `wpa`, `wpa2` and `wpa3`
 
-Create this file with your SSID information in it and then run the makefile `make` target, and it will produce a
-`porky_w` binary with that SSID as the default SSID. Run `make uf2` to generate a UF2 file for download.
+Create this file with your SSID information in it and then run the makefile `make build-w` or `make build-w2` targets,
+and it will produce a `porky` binary with that SSID as the default SSID. Run `make uf2` to generate a UF2 file for
+download.
 
 ## Running Directly
 
 NOTE: Running directly requires a debug probe attached to your device and host, and `probe-rs` installed.
 
 If you have a debug probe for the Pico, and it is connected and also the Pico is connected to
-USB then you can use probe-rs to download and debug `porky_w`.
+USB then you can use probe-rs to download and debug `porky`.
+
+NOTE: `probe-rs` still doesn't have support for Pi Pico 2 (it's being worked on) so temporarily we have fallen back
+to using `picotool` to download the firmware to flash and run it. It doesn't support SWG debug yet. So, you will have to
+start the device with BOOTSEL, so it's waiting for a firmware fownload.
 
 [config.toml](./.cargo/config.toml) is where the runner for cargo is set up.
 
-You can use `make run` (which uses `cargo run --release`) to run `porky_w` directly.
+You can use `make run` or `make run-w` or `make run2` or `make run-w2` (which uses `cargo run --release`) to run `porky`
+directly.
 
 The release binary is about half the size of the debug binary, so downloading to the Pico W is much faster.
 
-The Pi Pico W will start running `porky_w` and you should start seeing log messages on the terminal where
-you are running `make run`. The boot process should end with `porky_w` connected to the configured (or default)
+The Pi Pico W will start running `porky` and you should start seeing log messages on the terminal where
+you are running `make run`. The boot process should end with `porky` connected to the configured (or default)
 Wi-Fi network, and the output of its IP and the port it is listening for TCP connections on.
 
 ## Creating a UF2 file
